+++ aurex/src/exercises/walking.rs
@@ -1,5 +1,5 @@
 pub struct WalkingBassExercise {
-    pub bpm: u64,
+    pub bpm: f64,
     pub loops: u32,
     pub root: wmidi::Note,
     pub chord_loop: Vec<Vec<wmidi::Note>>,
