+++ aurex/src/metronome/beat.rs
@@ -5,7 +5,7 @@ use super::Metronome;
 pub struct BeatMetronome {}
 
 impl Metronome for BeatMetronome {
-    fn generate(bpm: u64) -> Sequence {
+    fn generate(bpm: f64) -> Sequence {
         drums::metronome(bpm)
     }
 }
