+++ aurex/src/player/playtwice.rs
@@ -27,7 +27,7 @@ impl<E: Exercise, M: Metronome> Exercise for PlayTwice<E, M> {
         self.exercise.instrument()
     }
 
-    fn bpm(&self) -> u64 {
+    fn bpm(&self) -> f64 {
         self.exercise.bpm()
     }
 }
